+++ student-info-system/schooladmin/templates/schooladmin/home_admin.html
@@ -35,7 +35,7 @@
     ></iframe>
 <iframe id="My Profile" class="tabcontent" title="My Profile" loading="lazy" src="{% url 'schooladmin:profile' %}"
     ></iframe>
-<iframe id="My Messages" class="tabcontent" title="My Messages" loading="lazy" src="{% url 'schooladmin:messages' %}"
+<iframe id="My Messages" class="tabcontent" title="My Messages" loading="lazy" src="{% url 'sis:messages' %}"
     ></iframe>
 </span>
 {% endblock %}
