+++ student-info-system/schooladmin/views.py
@@ -1117,28 +1117,8 @@ def profile(request):
     data = {
         'user': the_user,
     }
-    data.update(
-        filtered_table2(
-            name='received',
-            qs=the_user.profile.sent_to.all(),
-            filter=FullReceivedMessageFilter,
-            table=MessageReceivedTable,
-            request=request,
-            wrap_list=False,
-            self_url=reverse('schooladmin:profile'),
-        ))
-    data.update(
-        filtered_table2(
-            name='sent',
-            qs=the_user.profile.sent_by.all(),
-            filter=FullSentMessageFilter,
-            table=MessageSentTable,
-            request=request,
-            wrap_list=False,
-            self_url=reverse('schooladmin:profile'),
-        ))
 
-    return render(request, 'schooladmin/profile.html', data)
+    return render(request, 'schooladmin/profile_admin.html', data)
 
 
 @role_login_required(Profile.ACCESS_ADMIN)
@@ -1146,119 +1126,3 @@ def profile_edit(request):
     return user_edit(request, request.user.id)
 
 
-@role_login_required(Profile.ACCESS_ADMIN)
-def usermessages(request):
-    the_user = request.user
-
-    sentFilter = FullSentMessageFilter
-    receivedFilter = FullReceivedMessageFilter
-    if the_user.profile.role == Profile.ACCESS_STUDENT:
-        sentFilter = SentMessageFilter
-        receivedFilter = ReceivedMessageFilter
-
-    data = {
-        'user': the_user,
-    }
-    data.update(
-        filtered_table(
-            name='received',
-            qs=the_user.profile.sent_to.all(),
-            filter=receivedFilter,
-            table=MessageReceivedTable,
-            request=request,
-            wrap_list=False,
-        ))
-    data.update(
-        filtered_table(
-            name='sent',
-            qs=the_user.profile.sent_by.all(),
-            filter=sentFilter,
-            table=MessageSentTable,
-            request=request,
-            wrap_list=False,
-        ))
-
-    return render(request, 'schooladmin/messages.html', data)
-
-
-@role_login_required(Profile.ACCESS_ADMIN)
-def message(request, id):
-    the_user = request.user
-    the_profile = the_user.profile
-    the_mess = Message.objects.get(id=id)
-    is_recipient = the_mess.recipient == the_profile
-    is_sender = the_mess.sender == the_profile
-    is_student = the_profile.role == Profile.ACCESS_STUDENT
-
-    if the_mess is None:
-        messages.error(request, 'Invalid message')
-        return redirect('schooladmin:messages')
-
-    if not (is_sender or is_recipient) and the_profile != Profile.ACCESS_ADMIN:
-        messages.error(request, 'Invalid message')
-
-    if request.method == 'POST':
-        if request.POST.get('unarchive', None) is not None:
-            messages.success(request, 'Message unarchived.')
-            the_mess.time_archived = None
-            the_mess.save()
-
-        elif request.POST.get('archive', None) is not None:
-            messages.success(request, 'Message archived.')
-            the_mess.time_archived = datetime.now(pytz.utc)
-            the_mess.save()
-
-        elif request.POST.get('reply', None) is not None:
-            messages.success(request, 'Trust me, you replied.')
-
-        elif request.POST.get('approvedrop', None) is not None:
-            the_student = Student.objects.get(id=the_mess.support_data['student'])
-            the_ssect = SectionStudent.objects.get(id=the_mess.support_data['section'])
-            the_student.drop_approved(sectionstudent=the_ssect, request_message=the_mess)
-            messages.success(request, 'Drop Approved.')
-
-        elif request.POST.get('rejectdrop', None) is not None:
-            the_mess.time_handled = datetime.now(pytz.utc)
-            the_mess.save()
-            messages.success(request, 'Trust me, you rejected the drop.')
-
-        elif request.POST.get('approvemajor', None) is not None:
-            the_student = Student.objects.get(id=the_mess.support_data['student'])
-            the_new_major = Major.objects.get(id=the_mess.support_data['major'])
-            the_student.major_change_approved(major=the_new_major, request_message=the_mess)
-            messages.success(request, 'Major Change Approved.')
-
-        elif request.POST.get('rejectmajor', None) is not None:
-            the_mess.time_handled = datetime.now(pytz.utc)
-            the_mess.save()
-            messages.success(request, 'Trust me, you rejected the major change.')
-
-        else:
-            messages.error(request, 'Something went wrong.')
-
-    # mark our received messages read. Don't touch sent messages.
-    if is_recipient and the_mess.time_read is None:
-        the_mess.time_read = datetime.now(pytz.utc)
-        the_mess.save()
-
-    handled = the_mess.time_handled is not None
-    handleable_message = the_mess.message_type in (Message.DROP_REQUEST_TYPE,
-                                                   Message.MAJOR_CHANGE_TYPE)
-    show_drop = is_recipient and the_mess.message_type == Message.DROP_REQUEST_TYPE \
-        and not handled
-    show_major = is_recipient and the_mess.message_type == Message.MAJOR_CHANGE_TYPE \
-        and not handled
-    return render(
-        request, 'schooladmin/message.html', {
-            'user': the_user,
-            'message': the_mess,
-            'show_approve_drop': show_drop,
-            'show_approve_major': show_major,
-            'show_archive': is_recipient,
-            'message_archived': the_mess.time_archived is not None,
-            'message_read': the_mess.time_read is not None,
-            'message_handled': handled,
-            'show_type': not is_student,
-            'show_handled': handleable_message and not is_student,
-            'show_read': True,
-        })
