+++ student-info-system/sis/urls.py
@@ -1,6 +1,5 @@
-from django.contrib import admin
-from django.contrib.auth import views as auth_views
-from django.urls import include, path
+from django.urls import path
+from sis.views import messages
 
 from . import views
 
@@ -8,4 +7,6 @@ app_name = 'sis'
 urlpatterns = [
     path('', views.index, name='index'),
     path('access_denied', views.access_denied, name='access_denied'),
+    path('messages', messages.usermessages, name='messages'),
+    path('message/<int:id>', messages.message, name='message'),
 ]
