+++ student-info-system/sis/views/__init__.py
@@ -1,7 +1,7 @@
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
-from .models import Profile
+from sis.models import Profile
 
 
 @login_required
